refactor persistence

+++ service-api/src/main/java/com/github/al/realworld/api/query/GetArticlesResult.java
@@ -34,6 +34,6 @@ import java.util.List;
 public class GetArticlesResult {
 
     private List<ArticleDto> articles;
-    private Integer articlesCount;
+    private Long articlesCount;
 
 }
