+++ service-api/src/main/java/com/github/al/realworld/api/query/GetFeedResult.java
@@ -34,6 +34,6 @@ import java.util.List;
 public class GetFeedResult {
 
     private List<ArticleDto> articles;
-    private Integer articlesCount;
+    private Long articlesCount;
 
 }
